贴图可选

+++ assets/code/shader/program.h
@@ -26,7 +26,7 @@ public:
 	#pragma region 传输基本数据类型
 	void setBool(const std::string& name, bool value) const
 	{
-		glUniform1i(glGetUniformLocation(ID, name.c_str()), (int)value);
+		glUniform1i(glGetUniformLocation(ID, name.c_str()), value);
 	}
 	
 	void setInt(const std::string& name, int value) const
